backend/requirements.txt
fastapi
uvicorn
python-dotenv
nltk
spotipy
pydantic
